Module Ppl

A shallowly embedded DSL for Probabilistic Programming

This is a library for universal probabilistic programming using distributions as monads. Includes several approximate inference algorithms. Contains utilities to plot distributions and evaluate correctness of inference.

Core

module Dist : sig ... end

Module used for defining probabilistic models

module Primitive : sig ... end

Module defining a type for primitive distributions

module Empirical : sig ... end

A module for empirical distributions generated from samplers

module Inference : sig ... end

Implementation of inference algorithms

Extra

module Plot : sig ... end

Plotting utilies

module Evaluation : sig ... end

A module for evaluating the correctness of models and inference procedures

module Helpers : sig ... end

Utilities for working with distributions