Module Ppl__.Evaluation

type 'a samples = 'a Ppl.Empirical.Discrete.t

type 'a dist = 'a Ppl.Dist.dist

KL-Divergence

val kl_discrete : ?⁠n:int -> 'a Ppl.Primitive.t -> 'a dist -> float

Find the KL divergence for two discrete distributions

val kl_continuous : ?⁠n:int -> float Ppl.Primitive.t -> float dist -> float

Find the KL divergence for two continuous distributions

val kl_cum_discrete : int array -> bool Ppl.Primitive.t -> bool dist -> (int * float) array

val kl_cum_continuous : int array -> float Ppl.Primitive.t -> float dist -> (int * float) array

Hypothesis Tests

val kolmogorov_smirnov : ?⁠n:int -> ?⁠alpha:float -> float dist -> float Ppl.Primitive.t -> Owl_stats.hypothesis

Perform kolmogorov smirnov test, returns a hypothesis which is true if the null hypothesis is rejected

val chi_sq : ?⁠n:int -> ?⁠alpha:float -> 'a dist -> 'a Ppl.Primitive.t -> Owl_stats.hypothesis

Perform chi-squared test, returns a hypothesis which is true if the null hypothesis is rejected